+++ src/app.html
@@ -6,10 +6,10 @@
     <!-- <link rel="icon" href="%svelte.assets%/favicon.png" /> -->
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/phi-jp/meltline@v0.1.13/meltline.css">
-    
-    %svelte.head%
+    <link rel="stylesheet" href="/icons/icons.css" rel="stylesheet">
+    %sveltekit.head%
   </head>
-  <body>
-    %svelte.body%
+  <body data-sveltekit-preload-data="hover">
+    <div style="display: contents">%sveltekit.body%</div>
   </body>
 </html>
