+++ src/routes/posts/[id]/+page.js
@@ -1,0 +1,8 @@
+export async function load({params, fetch}) {
+  let res = await fetch(`https://jsonplaceholder.typicode.com/posts/${params.id}`);
+  let json = await res.json();
+
+  return {
+    post: json,
+  };
+}
