+++ src/routes/posts/[id]/+page.svelte
@@ -1,0 +1,16 @@
+<script>
+  import { page } from '$app/stores';
+  import { Meta } from '$lib/index';
+  $: ({ post } = $page.data);
+</script>
+
+<Meta title='{post.id}. {post.title}' description='{post.body}' />
+
+<template lang='pug'>
+  main.container-640.px16.py32
+    h1.mb32 {post.id}. {post.title}
+    p.mb32 {post.body}
+
+    a(sveltekit:prefetch, href='/') top
+
+</template>
