+++ svelte.config.js
@@ -1,32 +1,20 @@
+import path from 'path';
 import adapter from '@sveltejs/adapter-node';
 import sveltePreprocess from 'svelte-preprocess';
 const preprocess = sveltePreprocess({
   typescript: true,
   // ...
 });
-import path from 'path';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
   kit: {
     adapter: adapter(),
 
-    // Override http methods in the Todo forms
-    methodOverride: {
-      allowed: ['PATCH', 'DELETE']
+    alias: {
+      'svelte-modal-manager': path.resolve('./src/lib'),
+      $components: path.resolve('./src/components'),
     },
-
-    vite: {
-      mode: process.env.NODE_ENV,
-      define: {
-        'process.env': process.env,
-      },
-      resolve: {
-        alias: {
-          $components: path.resolve('./src/components')
-        }
-      }
-    }
   },
   preprocess,
 };
